fix: dashboard constants moved
