+++ packages/constants/src/index.ts
@@ -25,3 +25,4 @@ export * from "./inbox";
 export * from "./profile";
 export * from "./workspace-drafts";
 export * from "./label";
+export * from "./dashboard";
