+++ web/core/components/dashboard/widgets/issue-panels/tabs-list.tsx
@@ -1,8 +1,8 @@
 import { observer } from "mobx-react";
 import { Tab } from "@headlessui/react";
-import { TIssuesListTypes } from "@plane/types";
 // helpers
-import { EDurationFilters, FILTERED_ISSUES_TABS_LIST, UNFILTERED_ISSUES_TABS_LIST } from "@/constants/dashboard";
+import { EDurationFilters, FILTERED_ISSUES_TABS_LIST, UNFILTERED_ISSUES_TABS_LIST } from "@plane/constants";
+import { TIssuesListTypes } from "@plane/types";
 import { cn } from "@/helpers/common.helper";
 // types
 // constants
